Modified and refactored the unit tests.
