+++ src/DwitTech.AccountService.WebApi/Controllers/UserController.cs
@@ -42,7 +42,11 @@ namespace DwitTech.AccountService.WebApi.Controllers
             try
             {
                 var loginResult = await _authenticationService.AuthenticateUserLogin(loginDetails.Email, loginDetails.Password);
-                return Ok(loginResult);
+                if (loginResult == null)
+                {
+                    return BadRequest("Invalid email or password.");
+                }
+                return Ok(new { message = "Login successful." });
             }
             catch (Exception e)
             {
